First commit of materialize table script and README.
The SQL creates a materialized table for CWV data from exported GA data.
This table will be used later on with reporting to save on compute
costs. The table names will be replaced by the install script when run,
as paramterized table names are not supported.

The README is a first draft with the steps for deploying the solution.

Change-Id: I9bfbdae2ae902e0abb47f3e0b113d705b36746ba