First commit of notification email cloud run function.
The current code fetches the CWV values from bigquery using the stored
procedure, and then compares the values to threshold values defined as
environment variables (to be set in the dockerfile). Test coverage seems
poor, but most of the code is working with bigquery or boilerplate for
the cloud run webservice.

Change-Id: I986f5e2ff03ec9ba17cca013f02323dacea40c77