First commit of install.sh and minor correction to SQL.
This is the first commit of the install script. Currently it only
deploys the SQL scripts. Deploying the cloud run funcion is to come.

The chnges to the sql files are correcting some autoformatting in the
template values for project and ga id, as well a making the template
values consistent.

Change-Id: Ia092b4b3645f8d27a4c401d1de5fc0730e8001cd