Completed install script, fixed bug in SQL, and added license.
Completed the installation script. It's now deploying the Cloud Run function
and creating the eventarc trigger. The code was split into functions and
cleaned up to match go/shellstyle.

Backticks were added to the materialized table name to ensure project names
with dashes are properly handled.

The license header was added where missing.

Change-Id: Ibdb79ed5f8ee7d65a5a6b4c4fe6a521f5e9053d6