Deploying SQL for materialized table now schedules the query too.
Updated the install script so that the query to materialize the CWV summary
data is scheduled every 24 hours. Had to cat the contents of the file due to bq
needing the user to paste a version_info string into the console. A redirect
results in the console getting an EOF and dying.

Changed all of the single-quotes to double in the SQL.

Added the license to the create_p75 SQL script.

Change-Id: I9a5ee86cdfb1480c585c07de8e143ac867605005