Adding all the permissions needed to deploy to the created role.
This adds all the permissions to the role and then adds the service
account to the role to ensure there aren't any problems.

Also added a 30s sleep after the transfer config is submitted, as the
summary table must be there to continue and this is taking longer than
in the past.